removes commented code, console logs and refactors select options function
