+++ src/components/Footer/CopyrightComponent.tsx
@@ -5,8 +5,10 @@ export default function CopyrightComponent() {
 	return (
 		<Typography variant="body2" color="text.secondary" align="center">
 			{'Copyright © '}
-			<Link color="inherit" href="https://mui.com/">
-				Your Website
+			<Link
+				color="inherit"
+				href="https://www.linkedin.com/in/vadim-kravchenko-07833674/">
+				Linkedin profile
 			</Link>{' '}
 			{new Date().getFullYear()}
 			{'.'}
