+++ src/components/Footer/FooterComponent.tsx
@@ -5,7 +5,6 @@ import CopyrightComponent from './CopyrightComponent';
 export default function FooterComponent() {
 	return (
 		<Box sx={{bgcolor: 'background.paper', p: 6}} component="footer">
-			{/* Footer */}
 			<Typography variant="h6" align="center" gutterBottom>
 				Footer
 			</Typography>
@@ -17,7 +16,6 @@ export default function FooterComponent() {
 				Something here to give the footer a purpose!
 			</Typography>
 			<CopyrightComponent />
-			{/* End footer */}
 		</Box>
 	);
 }
